requirements.txt
streamlit==1.31.0
openai==1.11.1
toml~=0.10.2
pandas~=2.1.3
tabulate==0.9.0
python-dotenv==1.0.1
pyperclip
pyautogen
requests
streamlit
openai
toml
pandas
tabulate
python-dotenv
httpx
asyncio~=3.4.3
pytest
fastapi
uvicorn
gunicorn
jupyter

